Speeding up update-db.sh
Importing large amounts of data with views and indices in place can be
very slow compared to importing data and then (re)applying indices and
views.

This change will orchestrate `fastupdate-db.sh` in such a way that firstly the
schema will be created, read, and all tables will be cascade-dropped, so
that dependant views and indices will be also dropped. Secondly, tables
will be imported, and lastly a `pg_restore` of the schema only will be
performed to (re)apply indices and views. This will have the side-effect
of some harmless messages (tables already exists).